+++ R/importExperiment.R
@@ -10,14 +10,20 @@
         "Intensity"
     } else if (patmatch1 == "iBAQ.") {
         "iBAQ"
+    } else if (patmatch1 == "Abundance.") {
+        "Abundance"
     } else if (patmatch1 == "Abundance.F[0-9]+.") {
         "Abundance"
     } else if (patmatch1 == "Abundance.F.+.Sample.") {
         "Abundance"
+    } else if (patmatch1 == "Abundances.Count.") {
+        "Abundances.count"
     } else if (patmatch1 == "Abundances.Count.F[0-9]+.") {
         "Abundances.count"
     } else if (patmatch1 == "Abundances.Count.F.+.Sample.") {
         "Abundances.count"
+    } else if (patmatch1 == "Abundances.Normalized.") {
+        "Abundances.normalized"
     } else if (patmatch1 == "Abundances.Normalized.F[0-9]+.") {
         "Abundances.normalized"
     } else if (patmatch1 == "Abundances.Normalized.F.+.Sample.") {
@@ -70,7 +76,7 @@
 #' @param iColPattern Character scalar defining a regular expression to
 #'     identify sample columns. For MaxQuant output, this is typically
 #'     one of "^iBAQ\\.", "^LFQ\\.intensity\\." or "^Intensity\\.". For PD,
-#'     it is typically "^Abundance\\.F[0-9]+\\." or
+#'     it is typically "^Abundance\\.", "^Abundance\\.F[0-9]+\\." or
 #'     "^Abundance\\.F.+\\.Sample\\.". For FragPipe,
 #'     it is typically "\\.MaxLFQ\\.Intensity$". Columns matching the
 #'     given pattern will form the first assay in the output object.
@@ -108,10 +114,13 @@ importExperiment <- function(inFile, iColPattern, includeOnlySamples = "",
         "^Unique\\.peptides\\.", "^Razor\\.+unique\\.peptides\\.",
         "^Peptides\\.", "^iBAQ\\.", "^Identification\\.type\\.",
         ## ProteomeDiscoverer
+        "^Abundance\\.",
         "^Abundance\\.F[0-9]+\\.",
         "^Abundance\\.F.+\\.Sample\\.",
+        "^Abundances\\.Count\\.",
         "^Abundances\\.Count\\.F[0-9]+\\.",
         "^Abundances\\.Count\\.F.+\\.Sample\\.",
+        "^Abundances\\.Normalized\\.",
         "^Abundances\\.Normalized\\.F[0-9]+\\.",
         "^Abundances\\.Normalized\\.F.+\\.Sample\\.",
         "^Abundances\\.Grouped\\.Count\\.",
@@ -148,8 +157,8 @@ importExperiment <- function(inFile, iColPattern, includeOnlySamples = "",
         stop("Specifying ", iColPattern, " as the main assay is currently ",
              "not supported.")
     }
-    ## The exception is Abundances.Grouped - in this case, allow it but give
-    ## a warning
+    ## The exception is Abundances.Grouped - in this case,
+    ## allow it but give a warning
     if (iColPattern %in% c("^Abundances\\.Grouped\\.",
                            "^Abundances.Grouped.")) {
         warning("Note that the specified iColPattern may match different ",
@@ -177,12 +186,28 @@ importExperiment <- function(inFile, iColPattern, includeOnlySamples = "",
     ## multiple patterns corresponding to the same assay)
     if (iColPattern %in% c("^Abundance\\.F[0-9]+\\.", "^Abundances\\.Count\\.F[0-9]+\\.",
                            "^Abundances\\.Normalized\\.F[0-9]+\\.")) {
-        pats <- pats[!(pats %in% c("^Abundance\\.F.+\\.Sample\\.",
+        pats <- pats[!(pats %in% c("^Abundance\\.Sample\\.",
+                                   "^Abundance\\.F.+\\.Sample\\.",
                                    "^Abundances\\.Count\\.F.+\\.Sample\\.",
-                                   "^Abundances\\.Normalized\\.F.+\\.Sample\\."))]
-    } else {
-        pats <- pats[!(pats %in% c("^Abundance\\.F[0-9]+\\.",
+                                   "^Abundances\\.Normalized\\.F.+\\.Sample\\.",
+                                   "^Abundance\\.",
+                                   "^Abundances\\.Count\\.",
+                                   "^Abundances\\.Normalized\\."))]
+    } else if (iColPattern %in% c("^Abundance\\.", "^Abundances\\.Count\\.",
+                                  "^Abundances\\.Normalized\\.")) {
+        pats <- pats[!(pats %in% c("^Abundance\\.Sample\\.",
+                                   "^Abundance\\.F.+\\.Sample\\.",
+                                   "^Abundances\\.Count\\.F.+\\.Sample\\.",
+                                   "^Abundances\\.Normalized\\.F.+\\.Sample\\.",
+                                   "^Abundance\\.F[0-9]+\\.",
                                    "^Abundances\\.Count\\.F[0-9]+\\.",
+                                   "^Abundances\\.Normalized\\.F[0-9]+\\."))]
+    } else {
+        pats <- pats[!(pats %in% c("^Abundance\\.",
+                                   "^Abundance\\.F[0-9]+\\.",
+                                   "^Abundances\\.Count\\.",
+                                   "^Abundances\\.Count\\.F[0-9]+\\.",
+                                   "^Abundances\\.Normalized\\.",
                                    "^Abundances\\.Normalized\\.F[0-9]+\\."))]
     }
 
@@ -248,7 +273,9 @@ importExperiment <- function(inFile, iColPattern, includeOnlySamples = "",
     if (any(duplicated(names(assayList)))) {
         ## Should never end up in here
         #nocov start
-        warning("Multiple column patterns corresponding to the same assay name")
+        warning("Multiple column patterns corresponding to the same assay name: ",
+                paste(names(assayList)[duplicated(names(assayList))],
+                      collapse = "; "))
         #nocov end
     }