+++ NEWS.md
@@ -3,6 +3,7 @@
 * Change approach for injecting values into the Rmd file, to avoid the need to duplicate escapes
 * Allow the user to set the column to colour by in the PCA
 * Include 'extra' columns in output from runPTMTest
+* Allow '^Abundance.' iColPattern
 
 # einprot 0.6.10